# documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cities API
  version: 1.0.0
  description: RESTful API for city data with spatial capabilities
  contact:
    name: Aaron Baggot
    programme: TU856
    Module: Web Mapping
    email: C22716399@mytudublin.ie
  license:
    name: MIT
paths:
  /api/cities/:
    get:
      operationId: cities_list
      description: Filter, search and paginate.
      summary: List cities
      parameters:
      - in: query
        name: country
        schema:
          type: string
      - in: query
        name: is_capital
        schema:
          type: boolean
      - in: query
        name: max_founded_year
        schema:
          type: integer
      - in: query
        name: max_population
        schema:
          type: integer
      - in: query
        name: min_founded_year
        schema:
          type: integer
      - in: query
        name: min_population
        schema:
          type: integer
      - in: query
        name: name
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: region
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - Cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCityListList'
          description: ''
    post:
      operationId: cities_create
      description: Filter, search and paginate.
      summary: List cities
      tags:
      - Cities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityCreate'
          description: ''
  /api/cities/{id}/:
    get:
      operationId: cities_retrieve
      description: Retrieve, update, or delete a city instance
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityDetail'
          description: ''
    put:
      operationId: cities_update
      description: Retrieve, update, or delete a city instance
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - cities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
        required: true
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityCreate'
          description: ''
    patch:
      operationId: cities_partial_update
      description: Retrieve, update, or delete a city instance
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - cities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCityCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCityCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCityCreateRequest'
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityCreate'
          description: ''
    delete:
      operationId: cities_destroy
      description: Retrieve, update, or delete a city instance
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '204':
          description: No response body
  /api/cities/bbox/:
    post:
      operationId: cities_bbox_create
      description: Find cities within a bounding box
      tags:
      - Spatial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoundingBoxRequest'
            examples:
              BBoxAroundIreland-ish:
                value:
                  min_latitude: 51.3
                  min_longitude: -10.6
                  max_latitude: 55.5
                  max_longitude: -5.4
                summary: BBox around Ireland-ish
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BoundingBoxRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BoundingBoxRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/cities/countries/:
    get:
      operationId: cities_countries_retrieve
      description: Return list of countries with city counts
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/cities/geojson/:
    get:
      operationId: cities_geojson_list
      description: Return cities as GeoJSON for mapping applications
      parameters:
      - in: query
        name: country
        schema:
          type: string
      - in: query
        name: is_capital
        schema:
          type: boolean
      - in: query
        name: max_founded_year
        schema:
          type: integer
      - in: query
        name: max_population
        schema:
          type: integer
      - in: query
        name: min_founded_year
        schema:
          type: integer
      - in: query
        name: min_population
        schema:
          type: integer
      - in: query
        name: name
        schema:
          type: string
      - in: query
        name: region
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityGeoJSONList'
          description: ''
  /api/cities/info/:
    get:
      operationId: cities_info_retrieve
      description: Return API information and available endpoints
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/cities/stats/:
    get:
      operationId: cities_stats_retrieve
      description: Return statistical summary of cities data
      tags:
      - cities
      security:
      - cookieAuth: []
      - tokenAuth: []
      - {}
      responses:
        '200':
          description: No response body
  /api/cities/within-radius/:
    post:
      operationId: cities_within_radius_create
      description: Find cities within specified radius of a point
      tags:
      - Spatial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistanceRequest'
            examples:
              FindCitiesWithin50kmOfDublin:
                value:
                  latitude: 53.3498
                  longitude: -6.2603
                  radius_km: 50
                summary: Find cities within 50km of Dublin
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DistanceRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DistanceRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
components:
  schemas:
    BoundingBoxRequest:
      type: object
      description: Serializer for bounding box queries
      properties:
        min_latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        min_longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
        max_latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        max_longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
      required:
      - max_latitude
      - max_longitude
      - min_latitude
      - min_longitude
    CityCreate:
      type: object
      description: Serializer for creating/updating cities
      properties:
        name:
          type: string
          maxLength: 200
        country:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 200
        population:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_capital:
          type: boolean
        founded_year:
          type: integer
          maximum: 2025
          minimum: -4000
          nullable: true
      required:
      - country
      - name
      - population
    CityCreateRequest:
      type: object
      description: Serializer for creating/updating cities
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: string
          minLength: 1
          maxLength: 100
        region:
          type: string
          maxLength: 200
        population:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_capital:
          type: boolean
        founded_year:
          type: integer
          maximum: 2025
          minimum: -4000
          nullable: true
        latitude:
          type: number
          format: double
          writeOnly: true
        longitude:
          type: number
          format: double
          writeOnly: true
      required:
      - country
      - latitude
      - longitude
      - name
      - population
    CityDetail:
      type: object
      description: Detailed serializer for individual city
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        country:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 200
        population:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_capital:
          type: boolean
        founded_year:
          type: integer
          maximum: 2025
          minimum: -4000
          nullable: true
        location:
          type: object
          properties:
            type:
              type: string
              enum:
              - Point
            coordinates:
              type: array
              items:
                type: number
                format: float
              example:
              - 12.9721
              - 77.5933
              minItems: 2
              maxItems: 3
          description: Geographic coordinates
        latitude:
          type: string
          readOnly: true
        longitude:
          type: string
          readOnly: true
      required:
      - country
      - id
      - latitude
      - location
      - longitude
      - name
      - population
    CityGeoJSON:
      type: object
      description: GeoJSON serializer for mapping applications
      properties:
        type:
          $ref: '#/components/schemas/GisFeatureEnum'
        id:
          type: integer
          readOnly: true
        geometry:
          type: object
          properties:
            type:
              type: string
              enum:
              - Point
            coordinates:
              type: array
              items:
                type: number
                format: float
              example:
              - 12.9721
              - 77.5933
              minItems: 2
              maxItems: 3
        properties:
          type: object
          properties:
            name:
              type: string
              maxLength: 200
            country:
              type: string
              maxLength: 100
            region:
              type: string
              maxLength: 200
            population:
              type: integer
              maximum: 2147483647
              minimum: 0
            is_capital:
              type: boolean
            founded_year:
              type: integer
              maximum: 2025
              minimum: -4000
              nullable: true
    CityGeoJSONList:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/GisFeatureCollectionEnum'
        features:
          type: array
          items:
            $ref: '#/components/schemas/CityGeoJSON'
    CityList:
      type: object
      description: Serializer for listing cities
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        country:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 200
        population:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_capital:
          type: boolean
        founded_year:
          type: integer
          maximum: 2025
          minimum: -4000
          nullable: true
        latitude:
          type: string
          readOnly: true
        longitude:
          type: string
          readOnly: true
      required:
      - country
      - id
      - latitude
      - longitude
      - name
      - population
    DistanceRequest:
      type: object
      description: Serializer for distance-based queries
      properties:
        latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
        radius_km:
          type: number
          format: double
          maximum: 20000
          minimum: 0.1
      required:
      - latitude
      - longitude
      - radius_km
    GisFeatureCollectionEnum:
      type: string
      enum:
      - FeatureCollection
    GisFeatureEnum:
      type: string
      enum:
      - Feature
    PaginatedCityListList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/CityList'
    PatchedCityCreateRequest:
      type: object
      description: Serializer for creating/updating cities
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        country:
          type: string
          minLength: 1
          maxLength: 100
        region:
          type: string
          maxLength: 200
        population:
          type: integer
          maximum: 2147483647
          minimum: 0
        is_capital:
          type: boolean
        founded_year:
          type: integer
          maximum: 2025
          minimum: -4000
          nullable: true
        latitude:
          type: number
          format: double
          writeOnly: true
        longitude:
          type: number
          format: double
          writeOnly: true
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"